completed

+++ DVD.cs
@@ -8,9 +8,9 @@ namespace csharp_biblioteca_db
 {
     public class DVD : Documento
     {
-        public int Durata { get; set; }
+        public long Durata { get; set; }
 
-        public DVD(int Codice, string Titolo, string Settore, int Durata, string Scaffale) : base(Codice, Titolo, Settore, Scaffale)
+        public DVD(long Codice, string Titolo, string Settore, int Durata, string Scaffale) : base(Codice, Titolo, Settore, Scaffale)
         {
             this.Durata = Durata;
         }
